requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
black==21.4b2
brotlipy==0.7.0
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.0.0
cycler==0.11.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@728ee8b0cfe045301a2ca479354b2fb969390e8e
docker-pycreds==0.4.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
fvcore==0.1.5.post20220512
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.1
hydra-core==1.1.2
idna==3.4
importlib-metadata==4.11.3
importlib-resources==5.2.3
iopath==0.1.9
kiwisolver==1.4.2
lmdb==1.3.0
Markdown==3.3.7
matplotlib==3.5.2
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.1.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.0.1
portalocker==2.4.0
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pyDeprecate==0.3.2
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.3
pytz==2022.6
PyYAML==6.0
regex==2022.4.24
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
toml==0.10.2
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.8.2
torchvision==0.12.0
tqdm==4.64.0
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.12.16
Werkzeug==2.1.2
yacs==0.1.8
yarl==1.7.2
zipp==3.8.0
